Game logic complete
Moved new game (+NEWGAME button) logic into app.js. Completed logic to
get user entry, test that it’s a number between 1 - 100. If it’s a
valid entry the user screen is updated and a comparison function is
called to test the entry against the generated random number and
provide the user with feedback regarding where their guess stands.